+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 
 ## Unreleased
 #### Added
-- Support for Buffer parameters and results
+- Support for Buffer and Stream parameters and results
+- Validates exposed API metadata such as `validate`, `validateResponse` or `responseSchema`
+- Usage of Yarn 
 
 #### Changed
 - Reformat CHANGELOG to follow [Keep a Changelog](https://keepachangelog.com) recommandations